Revert "测速"
This reverts commit 281797ad049f98963cee2702a4376474c49f2e3f.

+++ backend/config/config.go
@@ -24,7 +24,7 @@ type Peer struct {
 	Port     uint16 `json:"port"`
 	Addr     string `json:"addr"`
 	UUID     string `json:"uuid"`
-	Ping     int64  `json:"ping"`
+	Ping     uint   `json:"ping"`
 }
 
 func (p *Peer) Domain() string {
